+++ app/src/main/java/io/ipoli/android/app/settings/SettingsActivity.java
@@ -40,9 +40,9 @@ import butterknife.OnClick;
 import io.ipoli.android.BuildConfig;
 import io.ipoli.android.Constants;
 import io.ipoli.android.R;
-import io.ipoli.android.app.AndroidCalendarEventParser;
+import io.ipoli.android.app.sync.AndroidCalendarEventParser;
 import io.ipoli.android.app.App;
-import io.ipoli.android.app.SyncAndroidCalendarProvider;
+import io.ipoli.android.app.sync.SyncAndroidCalendarProvider;
 import io.ipoli.android.app.TimeOfDay;
 import io.ipoli.android.app.activities.BaseActivity;
 import io.ipoli.android.app.events.EventSource;
