+++ app/src/main/java/io/ipoli/android/app/sync/AndroidCalendarLoader.java
@@ -13,9 +13,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-import io.ipoli.android.app.AndroidCalendarEventParser;
-import io.ipoli.android.app.InstanceData;
-import io.ipoli.android.app.SyncAndroidCalendarProvider;
 import io.ipoli.android.app.persistence.CalendarPersistenceService;
 import io.ipoli.android.app.utils.DateUtils;
 import io.ipoli.android.app.utils.LocalStorage;
