+++ app/src/main/java/io/ipoli/android/app/sync/InstanceData.java
@@ -1,4 +1,4 @@
-package io.ipoli.android.app;
+package io.ipoli.android.app.sync;
 
 public class InstanceData {
