+++ app/src/main/java/io/ipoli/android/app/sync/SyncAndroidCalendarProvider.java
@@ -1,4 +1,4 @@
-package io.ipoli.android.app;
+package io.ipoli.android.app.sync;
 
 import android.content.ContentResolver;
 import android.content.ContentUris;
