+++ app/src/main/java/io/ipoli/android/store/fragments/UpgradeStoreFragment.java
@@ -30,10 +30,10 @@ import butterknife.ButterKnife;
 import butterknife.Unbinder;
 import io.ipoli.android.Constants;
 import io.ipoli.android.R;
-import io.ipoli.android.app.AndroidCalendarEventParser;
+import io.ipoli.android.app.sync.AndroidCalendarEventParser;
 import io.ipoli.android.app.App;
 import io.ipoli.android.app.BaseFragment;
-import io.ipoli.android.app.SyncAndroidCalendarProvider;
+import io.ipoli.android.app.sync.SyncAndroidCalendarProvider;
 import io.ipoli.android.app.events.EventSource;
 import io.ipoli.android.app.events.ScreenShownEvent;
 import io.ipoli.android.app.events.SyncCalendarRequestEvent;
